feature/css-modular-context add fully configurable component by using scss variables
