+++ src/components/features/decoupled-components/button/button.module.scss
@@ -1,0 +1,20 @@
+@use '../../../../styles/theme' as theme;
+
+$button-background: theme.$brand-main !default;
+$button-background-hover: theme.$brand-main-hover !default;
+$button-color: #fff !default;
+$button-font-family: theme.$brand-font-family !default;
+$button-font-size: 14px !default;
+$button-padding: 14px 20px;
+
+.button {
+  background-color: $button-background;
+  color: $button-color;
+  font-family: $button-font-family;
+  font-size: $button-font-size;
+  padding: $button-padding;
+
+  &:hover {
+    background-color: $button-background-hover;
+  }
+}
