+++ src/components/features/decoupled-components/hero-banner-component/hero-banner-component.module.scss
@@ -1,5 +1,5 @@
 $secondary-color: red;
 
 .custom_button_class button {
-  background-color: $secondary-color;
+  // background-color: $secondary-color;
 }
