+++ src/styles/theme.scss
@@ -1,0 +1,3 @@
+$brand-main: #ed6c25; // this could also be figma tokens later on
+$brand-main-hover: #f0854c;
+$brand-font-family: Lexend, sans-serif;
